Add draft of branching support.
The code tracks branch opcode that are successfully happening by adding two IR opcodes before and after each branch instruction. Then the normal hijack code function detect the branch triggering and can launch a special machinery that will create a set of temporary interleaving coping with the extra calls that have to be made while retaining the initial interleaving that led to the branch (for reproducability).